+++ src/briscas/__init__.py
@@ -1,3 +1,3 @@
 """Briscas card game."""
 
-__version__ = "2.1.7"
+__version__ = "2.2.0"